chore: Base tag information on script input rather than branch name
The container image tag information are now solely defined by the script input.
Due to the fact that branch information are not available during the release
process (because it indicated by a tag), it make sense to rather base the tag
determination on the stageParam, which is set i the respective pipeline step.

This also allows to further simplify the script assembling that tag and pushing
the images. Alongside, some variables have been renamed to clarify the meaning
of their values.
The pipelines uses 7 characters while the script only used 6 (second param of
substring behaves exclusive). This has been aligned.